requirements.txt
django==3.2.5
djangorestframework==3.13.1
django-cors-headers==3.11.0